+++ HISTORY.md
@@ -86,6 +86,7 @@
 
 ### 2023-10-20
 * lists: refactored some helper functions
+* threads: added VariableFrame >> branch()
 
 ### 2023-10-19
 * gui, extensions: added user edit event for language change